+++ pythonhere/window_here.py
@@ -5,13 +5,15 @@ from pathlib import Path
 import time
 
 from kivy.app import App
-from kivy.core.window import Window
 from kivy.lang import Builder
 from kivy.uix.boxlayout import BoxLayout
 
 
 def reset_window_environment() -> BoxLayout:
     """Remove PythonHere app widgets and styles."""
+    # import Window inside function to avoid early loading of the app config
+    from kivy.core.window import Window  # pylint: disable=import-outside-toplevel
+
     for widget in Window.children:
         widget.clear_widgets()
         Window.remove_widget(widget)
@@ -30,6 +32,8 @@ def unload_app_kv_styles():
 
 def load_kv_string(code: str, clear_style: bool):
     """Insert given rules into the Kivy Language Builder."""
+    from kivy.core.window import Window  # pylint: disable=import-outside-toplevel
+
     app = App.get_running_app()
 
     if clear_style:
@@ -50,6 +54,8 @@ def load_kv_string(code: str, clear_style: bool):
 
 def encoded_screenshot() -> str:
     """Return base64 encoded displayed image."""
+    from kivy.core.window import Window  # pylint: disable=import-outside-toplevel
+
     path = str(Path(f"screenshot_{time.time()}.png").resolve())
     Window.children[0].export_to_png(path)
     with open(path, "rb") as png_file: